server/requirements.txt
alembic==0.9.6
bcrypt==3.1.4
click==6.7
coverage==4.4.2
eventlet==0.31.0
flask==0.12.4
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Migrate==2.1.1
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.7.1
gunicorn==19.7.1
itsdangerous==0.24
Jinja2>=2.10.1
jsonschema==2.6.0
Mako==1.0.7
psycopg2-binary==2.9.1
pycparser==2.18
PyJWT==2.0.1
python-dateutil==2.6.1
pytz==2017.3
selenium==3.8.1
six==1.11.0
SQLAlchemy>=1.3.0
Werkzeug==0.15.5
python-dotenv==0.18.0
ibm-watson==5.2.1
ibm-cloud-sdk-core==3.10.1
black==19.10b0